+++ packages/react-form/src/hooks/list/test/list.test.tsx
@@ -60,6 +60,57 @@ describe('useList', () => {
     });
   });
 
+  it('accepts a list as argument', () => {
+    function ListArgTestList({list}: {list: Variant[]}) {
+      const variants = useList<Variant>(list);
+
+      return (
+        <ul>
+          {variants.map((fields, index) => (
+            <li key={index}>
+              <TextField
+                label="price"
+                name={`price${index}`}
+                {...fields.price}
+              />
+              <TextField
+                label="option"
+                name={`option${index}`}
+                {...fields.optionName}
+              />
+              <TextField
+                label="value"
+                name={`value${index}`}
+                {...fields.optionValue}
+              />
+            </li>
+          ))}
+        </ul>
+      );
+    }
+
+    const variants = randomVariants(4);
+    const wrapper = mount(<ListArgTestList list={variants} />);
+
+    variants.forEach(({price, optionName, optionValue}) => {
+      expect(wrapper).toContainReactComponent(TextField, {
+        value: price,
+        onChange: expect.any(Function),
+        onBlur: expect.any(Function),
+      });
+      expect(wrapper).toContainReactComponent(TextField, {
+        value: optionName,
+        onChange: expect.any(Function),
+        onBlur: expect.any(Function),
+      });
+      expect(wrapper).toContainReactComponent(TextField, {
+        value: optionValue,
+        onChange: expect.any(Function),
+        onBlur: expect.any(Function),
+      });
+    });
+  });
+
   describe('handlers', () => {
     describe('#onChange', () => {
       it('updates the value of the specific field and item when called with a value', () => {