+++ packages/react-form/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## Fixed
+
+- Overloaded `useList` to accept list as argument ([#1308](https://github.com/Shopify/quilt/issues/1308))
+
 ## [0.4.0] - 2020-03-04
 
 ### Added